+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Linux launch script
+- Download speed will now be displayed based on the average of download speeds in the last 2 seconds
 
 ### Fixed
 - Invalid display of progress value when unarchiving